pyproject.toml
[project]
name = "pydantic-ai-tool-budget"
version = "0.1.0"
description = "Per-tool budget reminders for Pydantic AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "sarth6" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = ["pydantic-ai-slim>=0.100.0"]

[project.urls]
Repository = "https://github.com/sarth6/pydantic-ai-tool-budget"
Issues = "https://github.com/sarth6/pydantic-ai-tool-budget/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_tool_budget"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-recording>=0.13.2",
    "pyright>=1.1",
    "ruff>=0.8",
    "pydantic-ai-slim[google]>=0.100.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]
